Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240809/합과 평균/sum-and-avg.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    double a, b;
+    int a, b;
     cin >> a >> b;
 
     cout.precision(2);
@@ -10,6 +10,6 @@ int main() {
     
     cout << a + b;
     cout << " ";
-    cout << (a + b) / 2;
+    cout << (double)(a + b) / 2;
     return 0;
 }
